requirements.txt
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pi==0.1.2
psycopg2==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
gunicorn==23.0.0
